+++ crates/openshell-cli/src/commands/mod.rs
@@ -3,3 +3,4 @@
 
 pub mod common;
 pub mod gateway;
+pub mod provider;
